+++ velox/vector/arrow/c/CMakeLists.txt
@@ -11,4 +11,11 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-add_subdirectory(c)
+add_library(velox_arrow_bridge Bridge.cpp)
+
+target_link_libraries(velox_arrow_bridge velox_memory velox_type velox_buffer
+                      velox_exception)
+
+if(${VELOX_BUILD_TESTING})
+  add_subdirectory(tests)
+endif()
